Bake upcoming schedule into prerendered HTML; add FAQ/WebSite schema
SEO: the event grid previously rendered only client-side, so crawlers
(especially JS-less ones like Yandex) saw an empty page.

- app/page.tsx is now a server component reading public/schedule.ics at
  build time and passing upcoming events to the renamed HomeClient
- useScheduleLoader seeds from initialEvents (no skeleton flash), keeps
  stale content on fetch failure instead of wiping it
- first render stays deterministic for hydration (UTC + scheduled status
  until mounted) while remaining fully crawlable
- JSON-LD additions: FAQPage (ru bundle) and WebSite/Organization
