+++ app/hooks/useScheduleLoader.ts
@@ -13,9 +13,11 @@ type ScheduleLoaderState = {
   reload: () => void;
 };
 
-export function useScheduleLoader(): ScheduleLoaderState {
-  const [events, setEvents] = useState<ScheduleEvent[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
+export function useScheduleLoader(initialEvents: ScheduleEvent[] = []): ScheduleLoaderState {
+  // Build-time prerendered schedule seeds the state so crawlers and JS-less
+  // visitors get real content immediately; the fetch below refreshes it.
+  const [events, setEvents] = useState<ScheduleEvent[]>(initialEvents);
+  const [isLoading, setIsLoading] = useState(initialEvents.length === 0);
   const [isError, setIsError] = useState(false);
   const [errorMessage, setErrorMessage] = useState('');
 
@@ -31,17 +33,17 @@ export function useScheduleLoader(): ScheduleLoaderState {
       }
       setEvents(parseSchedule(await response.text()));
     } catch (error) {
-      setEvents([]);
+      if (initialEvents.length === 0) {
+        setEvents([]);
+      }
       setIsError(true);
       setErrorMessage(error instanceof Error ? error.message : String(error));
     } finally {
       setIsLoading(false);
     }
-  }, []);
+  }, [initialEvents]);
 
   useEffect(() => {
-    // Deferred so the loader's state updates never run synchronously inside the
-    // effect body (react-hooks/set-state-in-effect).
     void Promise.resolve().then(load);
   }, [load]);
