+++ app/schema-org.tsx
@@ -1,35 +1,30 @@
-import { readFileSync } from 'node:fs';
-import path from 'node:path';
-import { DateTime } from 'luxon';
+import { DEFAULT_LANGUAGE, LANGUAGE_DEFINITIONS } from '../lib/language';
+import { buildFaqJsonLd, buildScheduleJsonLd, buildWebSiteJsonLd } from '../lib/schema-org';
+import { loadUpcomingEvents } from './load-schedule';
 
-import { parseSchedule, type ScheduleEvent } from '../lib/ics';
-import { buildScheduleJsonLd } from '../lib/schema-org';
+function JsonLd({ data }: { data: object }) {
+  return <script type="application/ld+json" dangerouslySetInnerHTML={{ __html: JSON.stringify(data) }} />;
+}
 
 /**
- * Server component executed once at build time: it reads the committed
- * public/schedule.ics and bakes schema.org JSON-LD into the prerendered
- * HTML so crawlers see the events without executing JavaScript.
+ * Server component executed once at build time: bakes schema.org structured
+ * data (upcoming events, FAQ, website identity) into the exported HTML so
+ * crawlers see it without executing JavaScript.
  */
 export function ScheduleJsonLd() {
-  let events: ScheduleEvent[] = [];
-  try {
-    const ics = readFileSync(path.join(process.cwd(), 'public', 'schedule.ics'), 'utf8');
-    const nowIso = DateTime.utc().toISO()!;
-    events = parseSchedule(ics)
-      .filter(event => event.startsAtUtc >= nowIso)
-      .sort((a, b) => a.startsAtUtc.localeCompare(b.startsAtUtc));
-  } catch (error) {
-    console.error('Failed to build schedule JSON-LD:', error);
-  }
+  const events = loadUpcomingEvents();
 
   if (events.length === 0) {
     return null;
   }
 
+  const faqItems = LANGUAGE_DEFINITIONS[DEFAULT_LANGUAGE].texts.faqItems;
+
   return (
-    <script
-      type="application/ld+json"
-      dangerouslySetInnerHTML={{ __html: JSON.stringify(buildScheduleJsonLd(events)) }}
-    />
+    <>
+      <JsonLd data={buildScheduleJsonLd(events)} />
+      <JsonLd data={buildFaqJsonLd(faqItems)} />
+      <JsonLd data={buildWebSiteJsonLd()} />
+    </>
   );
 }
