+++ lib/schema-org.ts
@@ -5,6 +5,8 @@ const MAX_SCHEMA_EVENTS = 50;
 
 type JsonLdObject = Record<string, unknown>;
 
+type FaqEntry = { question: string; answer: string };
+
 /** Builds a schema.org ItemList of upcoming SportsEvents for search-engine rich results. */
 export function buildScheduleJsonLd(events: ScheduleEvent[]): JsonLdObject {
   const itemListElement = events.slice(0, MAX_SCHEMA_EVENTS).map((event, index) => ({
@@ -29,5 +31,29 @@ export function buildScheduleJsonLd(events: ScheduleEvent[]): JsonLdObject {
     '@context': 'https://schema.org',
     '@type': 'ItemList',
     itemListElement,
+  };
+}
+
+export function buildFaqJsonLd(items: FaqEntry[]): JsonLdObject {
+  return {
+    '@context': 'https://schema.org',
+    '@type': 'FAQPage',
+    mainEntity: items.map(item => ({
+      '@type': 'Question',
+      name: item.question,
+      acceptedAnswer: { '@type': 'Answer', text: item.answer },
+    })),
+  };
+}
+
+export function buildWebSiteJsonLd(): JsonLdObject {
+  return {
+    '@context': 'https://schema.org',
+    '@type': 'WebSite',
+    name: 'RaceSync',
+    url: `${SITE_URL}/`,
+    inLanguage: 'ru',
+    description: 'F1, F2, F3 и MotoGP: расписание квалификаций, спринтов и гонок в вашем часовом поясе.',
+    publisher: { '@type': 'Organization', name: 'RaceSync', url: `${SITE_URL}/` },
   };
 }
